requirements.txt
boto3
markdown2
gunicorn
requests
psycopg2>=2.7,<3.0
Django>=2.0,<3.0
Django-admin-rangefilter
Django-allauth
Django-anymail
Django-autoslug
Django-braces
Django-ckeditor-filebrowser-filer
Django-cms
Django-cors-headers
Django-crispy-forms
Django-environ
Django-filer
Django-floppyforms
Django-impersonate
Django-import-export
Django-json-widget
Django-model-utils
Django-redis
Django-resized
Django-static-sitemaps
Django-storages
Django-tinymce
Djangocms-text-ckeditor
Djangorestframework
